Show image label in the lightbox
The sizing of the labels is not optimal in certain (hopefully only
edge-)cases, it can be improved later. When there is not enough space
for the label beneath the image, it would be better to move the image
higher and, if that is not enough, make the image a bit smaller.