Add image lightbox
* Add files&references for new lightbox functionality

* Remove hardcoded conflicting width attrib from imgs

* Remove conflicting super-general img width rule

It defines max-width 100% for all images on the `img` element level.
That, of course, conflics with zooming the lightboxed image. And I
didn't find a case where the rule would not be covered by some
lower-level more specific rule, so I'm fairly confident it is fine to
remove it.

* Show image label in the lightbox

The sizing of the labels is not optimal in certain (hopefully only
edge-)cases, it can be improved later. When there is not enough space
for the label beneath the image, it would be better to move the image
higher and, if that is not enough, make the image a bit smaller.

* Don't subtract padding from width

It behaves differently here than in my independent local test enviro…

* Responsive max-width of non-lightboxed images

* Move close button beyond lightbox bounding box

* fix label width and y-axis scroll

- The label <p> element inherited max-width from a more generic rule
  which made it too narrow as it overrode the width: 100% rule.
- Made the overflow: scroll rule specific to the y-axis to prevent some
  chromium-based browsers display an unneeded horizontal scrollbar

* Adjust img width breakpoints to match the template

* Add in-code documentation