Prevent executing pan funct on not zoomed image
Previously, when user tried to pan (grab by mouse) an image that was
scaled to the lightbox, no UX error occured byt the console showed an
error. To fix this, panning has to be prevented when
currentZoomScale==1. The impl. of this fix is ugly, though;
previously, the currentZoomScale var was neatly passed across the various functions,
but since some of the panning f~s are callbacks on elements and, hence,
a bit painful to be given attributes, I chose to make the variable global.