Improve panning by mouse behavior
In other words, introduce some dark magic calculations that normalize
the movement of the image being panned by mouse / finger. The gros of
the change is that when the image is zoomed only a little, the movement
needs to be faster than it was before, and vice versa, a more zoomed in
image needs to move slower in relation to the mouse/finger movement.